backend/requirements.txt
Django==3.1.4
django-cors-headers==3.8.0
djangorestframework==3.11.1
jsonfield==3.1.0